Connect with Friten on all Devices , Resuming Every Chat Right Where You Left Off
Friten is a social media app that allows users to connect, share, and engage with others through the creation and posting of fleets. Similar to tweets, fleets are short-lived messages that enable users to express their thoughts, ideas, and experiences. Alongside fleets, Friten offers various features such as following other users, discovering trending fleets, and interacting with followers through likes, comments, and direct messages. With its user-friendly interface and diverse community, Friten provides a platform for individuals to express themselves and connect with others in a convenient and engaging manner.